More clarity + flexibility for --gcs-dest
While --gcs-dest did already work, it previously added a slash to the end of
the specified URL prefix, even if the user already ended the parameter with one.
Refrain from doubling a single terminal slash. Also clarify the nature
of the parameter from URL to URL prefix; it is an URL and a prefix, but the
final URL will have a slash added if necessary followed by the image tarball
name.